# openapi/atuin-openapi.yaml
openapi: "3.0.0"
info:
  version: 0.1.0
  title: Atuin API
paths:
...
  /account:
    delete:
      summary: Delete account
      x-eov-operation-handler: AccountController
      operationId: deleteAccount
      tags:
        - account
      security:
        - tokenAuth: [ ]
      responses:
        '200':
          description: Expected response to a valid request
          content:
            text/plain:
              schema:
                type: string
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForbiddenError"
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundError"
        '500':
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalServerError"
...
  /account/tokens:
    get:
      summary: Get current user's tokens
      x-eov-operation-handler: AccountController
      operationId: getTokens
      tags:
        - account
      security:
        - tokenAuth: [ ]
      responses:
        '200':
          description: Expected response to a valid request
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  required:
                    - token
                    - expiresAt
                    - lastUsedAt
                  properties:
                    token:
                      type: string
                    expiresAt:
                      type: string
                      format: date-time
                    lastUsedAt:
                      type: string
                      format: date-time
                      nullable: true
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForbiddenError"
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundError"
        '500':
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalServerError"
  /account/tokens/{token}:
    parameters:
      - name: token
        in: path
        required: true
        schema:
          type: string
    delete:
      summary: Delete account's token
      x-eov-operation-handler: AccountController
      operationId: deleteToken
      tags:
        - account
      security:
        - tokenAuth: [ ]
      responses:
        '200':
          description: Expected response to a valid request
        '400':
          description: Invalid payload
          content:
            application/json:
              schema:
                  $ref: "#/components/schemas/InvalidPayloadError"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForbiddenError"
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundError"
        '500':
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalServerError"
  /account/validate:
    post:
      summary: Validate account
      x-eov-operation-handler: AccountController
      operationId: validateCode
      tags:
        - account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - code
                - username
                - password
              properties:
                username:
                  $ref: "#/components/schemas/NotNullOrEmptyString"
                password:
                  $ref: "#/components/schemas/NotNullOrEmptyString"
                code:
                  $ref: "#/components/schemas/NotNullOrEmptyString"
      responses:
        '200':
          description: Expected response to a valid request
          content:
            application/json:
              schema:
                type: object
        '400':
          description: Invalid payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvalidPayloadError"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForbiddenError"
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundError"
        '500':
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalServerError"
  /account/forgot-password:
    post:
      summary: Reset password account
      x-eov-operation-handler: AccountController
      operationId: forgotPassword
      tags:
        - account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - email
              properties:
                username:
                  $ref: "#/components/schemas/NotNullOrEmptyString"
                email:
                  type: string
                  format: email
      responses:
        '200':
          description: Expected response to a valid request
          content:
            application/json:
              schema:
                type: object
        '400':
          description: Invalid payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvalidPayloadError"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForbiddenError"
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundError"
        '500':
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalServerError"
  /account/reset-password:
    post:
      summary: Reset password account
      x-eov-operation-handler: AccountController
      operationId: resetPassword
      tags:
        - account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - code
                - password
              properties:
                code:
                  $ref: "#/components/schemas/NotNullOrEmptyString"
                password:
                  type: string
      responses:
        '200':
          description: Expected response to a valid request
          content:
            application/json:
              schema:
                type: object
        '400':
          description: Invalid payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvalidPayloadError"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForbiddenError"
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundError"
        '500':
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalServerError"
  /account/password:
    patch:
      summary: Change password
      x-eov-operation-handler: AccountController
      operationId: changePassword
      tags:
        - account
      security:
        - tokenAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - current_password
                - new_password
              properties:
                current_password:
                  $ref: "#/components/schemas/NotNullOrEmptyString"
                new_password:
                  $ref: "#/components/schemas/NotNullOrEmptyString"
      responses:
        '200':
          description: Expected response to a valid request
        '400':
          description: Invalid payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvalidPayloadError"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForbiddenError"
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundError"
        '500':
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalServerError"
  /account/{username}:
    parameters:
      - name: username
        in: path
        required: true
        schema:
          type: string
        description: The username of the account
    get:
      summary: Get user
      x-eov-operation-handler: AccountController
      operationId: getAccount
      tags:
        - account
      security:
        - tokenAuth: [ ]
      responses:
        '200':
          description: Expected response to a valid request
          content:
            application/json:
              schema:
                type: object
                properties:
                  username:
                    type: string
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForbiddenError"
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundError"
        '500':
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalServerError"
    delete:
      summary: Delete account
      x-eov-operation-handler: AccountController
      operationId: deleteAccount
      tags:
        - account
      security:
        - tokenAuth: [ ]
      responses:
        '200':
          description: Expected response to a valid request
          content:
            application/json:
              schema:
                type: object
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForbiddenError"
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundError"
        '500':
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalServerError"
  /account/{username}/{action}:
    parameters:
      - name: username
        in: path
        required: true
        schema:
          type: string
        description: The username of the account
      - name: action
        in: path
        required: true
        schema:
          type: string
          enum:
            - block
            - unblock
        description: The action to perform
    patch:
      summary: Block/unblock user
      x-eov-operation-handler: AccountController
      operationId: blockUnblockUser
      tags:
        - account
      security:
        - tokenAuth: [ ]
      responses:
        '200':
          description: Expected response to a valid request
          content:
            application/json:
              schema:
                type: object
                properties:
                  username:
                    type: string
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForbiddenError"
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundError"
        '500':
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalServerError"
    delete:
      summary: Delete account
      x-eov-operation-handler: AccountController
      operationId: deleteAccount
      tags:
        - account
      security:
        - tokenAuth: [ ]
      responses:
        '200':
          description: Expected response to a valid request
          content:
            application/json:
              schema:
                type: object
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForbiddenError"
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundError"
        '500':
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalServerError"
...
components:
  securitySchemes:
    tokenAuth:
      type: http
      scheme: Token
  schemas:
...
    NotNullOrEmptyString:
      type: string
      nullable: false
      minLength: 1
      pattern: '^\S+[\S ]*\S+$'
    BasicErrorModel:
      type: object
      required:
        - status
        - reason
      properties:
        status:
          type: integer
          format: int32
        reason:
          type: string
    InvalidPayloadError:
      allOf:
        - $ref: '#/components/schemas/BasicErrorModel'
        - type: object
          properties:
            status:
              example: 400
            reason:
              example: Invalid payload
    UnauthorizedError:
      allOf:
        - $ref: '#/components/schemas/BasicErrorModel'
        - type: object
          properties:
            status:
              example: 401
            reason:
              example: Unauthorized
    ForbiddenError:
      allOf:
        - $ref: '#/components/schemas/BasicErrorModel'
        - type: object
          properties:
            status:
              example: 403
            reason:
              example: Forbidden
    NotFoundError:
      allOf:
        - $ref: '#/components/schemas/BasicErrorModel'
        - type: object
          properties:
            status:
              example: 404
            reason:
              example: Not Found
...
    InternalServerError:
      allOf:
        - $ref: '#/components/schemas/BasicErrorModel'
        - type: object
          properties:
            status:
              example: 500
            reason:
              example: Internal Server Error